Resolve gcc-4.4 and 64-bit warnings. (dm)
git-svn-id: svn://mielke.cc/main/brltty@4543 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/ctbtest.c
@@ -96,7 +96,7 @@ static void
 writeUtf8Braille (FILE *stream, unsigned char cell) {
   Utf8Buffer utf8;
   size_t utfs = convertWcharToUtf8(cell|UNICODE_BRAILLE_ROW, utf8);
-  fprintf(stream, "%.*s", utfs, utf8);
+  fprintf(stream, "%.*s", (int)utfs, utf8);
 }
 
 typedef struct {
